+++ palladian-retrieval/src/main/java/ws/palladian/retrieval/search/videos/YouTubeSearcher.java
@@ -215,14 +215,14 @@ public final class YouTubeSearcher extends WebSearcher<WebVideoResult> {
     }
 
     @Override
-    public int getTotalResultCount(String query, Language language) throws SearcherException {
-        int hitCount = 0;
+    public long getTotalResultCount(String query, Language language) throws SearcherException {
+        long hitCount = 0;
         try {
             HttpResult httpResult = retriever.httpGet(getRequestUrl(query, 1, language));
             JSONObject root = new JSONObject(HttpHelper.getStringContent(httpResult));
             TOTAL_REQUEST_COUNT.incrementAndGet();
 
-            hitCount = root.getJSONObject("feed").getJSONObject("openSearch$totalResults").getInt("$t");
+            hitCount = root.getJSONObject("feed").getJSONObject("openSearch$totalResults").getLong("$t");
 
         } catch (JSONException e) {
             throw new SearcherException("Exception parsing the JSON response while searching for \"" + query
